+++ src/controllers/CartController.php
@@ -504,8 +504,6 @@ class CartController extends BaseFrontEndController
         $estimatedShippingAddress = $this->request->getParam('estimatedShippingAddress');
         $billingAddress = $this->request->getParam('billingAddress');
         $estimatedBillingAddress = $this->request->getParam('estimatedBillingAddress');
-        $fieldsLocationShippingAddress = $this->request->getParam('shippingAddress.fieldsLocation') ?? 'shippingAddress.fields';
-        $fieldsLocationBillingAddress = $this->request->getParam('billingAddress.fieldsLocation') ?? 'billingAddress.fields';
 
         // Use an address ID from the customer address book to populate the address
         $shippingAddressId = $this->request->getParam('shippingAddressId');
@@ -527,6 +525,10 @@ class CartController extends BaseFrontEndController
         } elseif ($shippingAddress && !$shippingIsBilling) {
             $this->_cart->sourceShippingAddressId = null;
             $this->_cart->setShippingAddress($shippingAddress);
+
+            if (!empty($shippingAddress['fields']) && $this->_cart->getShippingAddress()) {
+                $this->_cart->getShippingAddress()->setFieldValues($shippingAddress['fields']);
+            }
         }
 
         // Billing address
@@ -545,6 +547,10 @@ class CartController extends BaseFrontEndController
         } elseif ($billingAddress && !$billingIsShipping) {
             $this->_cart->sourceBillingAddressId = null;
             $this->_cart->setBillingAddress($billingAddress);
+
+            if (!empty($billingAddress['fields']) && $this->_cart->getBillingAddress()) {
+                $this->_cart->getBillingAddress()->setFieldValues($billingAddress['fields']);
+            }
         }
 
         // Estimated Shipping Address
@@ -595,9 +601,5 @@ class CartController extends BaseFrontEndController
             $this->_cart->sourceShippingAddressId = $this->_cart->sourceBillingAddressId;
             $this->_cart->shippingAddressId = $billingAddressId;
         }
-
-        // Custom fields
-        $this->_cart->getShippingAddress()->setFieldValuesFromRequest($fieldsLocationShippingAddress);
-        $this->_cart->getBillingAddress()->setFieldValuesFromRequest($fieldsLocationBillingAddress);
     }
 }