+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release Notes for Craft Commerce
 
-## Unreleased - 2022-05-06
+## Unreleased
+
+### Changed
+- Example templates address forms now show plain text custom fields.
 
 ### Fixed
 - Fixed a PHP error that occurred when making a payment on the Edit Order page. ([#2795](https://github.com/craftcms/commerce/issues/2795))